requirements.txt
anthropic>=0.58.0
lark-oapi>=1.4.14
PyYAML>=6.0.2
httpx>=0.27.0
pydantic>=2.9.0
numpy>=1.26.0
html2text>=2024.2.26
lxml>=5.2.0
